+++ app/Services/UserService.php
@@ -162,6 +162,29 @@ class UserService
 		return $logs;
 	}
 
+	public function searchPaginatedAll(int $page = 1, string $url, string $search): ?LengthAwarePaginator {
+		$fields = User::SELECT_FIELDS;
+
+		$query = self::getSearchQuery($fields);
+		$query = $query->where('username', 'LIKE', "%{$search}%")
+		               ->orWhere('email', 'LIKE', "%{$search}%");
+
+		if (Helper::env_bool('DEBUG_SEARCH_SQL')) {
+			$this->logger->info(self::getSqlFromQuery($query));
+		}
+
+		try {
+			$logs = $query
+				->paginate($this->items_per_page, $fields, 'page', $page)
+				->withPath($url);
+		} catch (\Exception $e) {
+			$this->logger->error("Query error: " . $e->getMessage() . PHP_EOL);
+			exit("Query error");
+		}
+
+		return $logs;
+	}
+
 	public function showPaginatedAliases(int $page = 1, string $url): ?LengthAwarePaginator {
 		$fields = User::SELECT_FIELDS;