+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2025-10-09
+- Search users
+
 ### 2025-10-08
 - Add reports by domain
 - showStats should not use full MailLogs results but counters only